requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
numpy==1.20.0
pandas==1.2.1
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-telegram-bot==13.2
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.23
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
